+++ ft_print_x.c
@@ -6,7 +6,7 @@
 /*   By: sdukic <sdukic@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/21 18:44:32 by sdukic            #+#    #+#             */
-/*   Updated: 2022/10/21 20:12:13 by sdukic           ###   ########.fr       */
+/*   Updated: 2022/10/21 20:59:10 by sdukic           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include "libprintf.h"
 #include "libft/libft.h"
 
-int	ft_print_x(unsigned int arg)
+int	ft_print_x(unsigned long int arg)
 {
 	ft_putnbr_base(arg, "0123456789abcdef");
 	return (ft_numlen_base(arg, 16));
